+++ Modules/Microsoft365DSC/Examples/Resources/EXOHostedOutboundSpamFilterPolicy/2-Update.ps1
@@ -33,9 +33,9 @@ Configuration Example
             BccSuspiciousOutboundMail                 = $False
             NotifyOutboundSpam                        = $False
             NotifyOutboundSpamRecipients              = @()
-            RecipientLimitExternalPerHour             = 0
-            RecipientLimitInternalPerHour             = 1 # Updated Property
-            RecipientLimitPerDay                      = 0
+            #RecipientLimitExternalPerHour             = 0
+            #RecipientLimitInternalPerHour             = 1 # Updated Property
+            #RecipientLimitPerDay                      = 0
             Ensure                                    = "Present"
             ApplicationId         = $ApplicationId
             TenantId              = $TenantId