+++ Modules/Microsoft365DSC/Examples/Resources/EXOHostedOutboundSpamFilterRule/3-Remove.ps1
@@ -21,7 +21,6 @@ Configuration Example
 
     Import-DscResource -ModuleName Microsoft365DSC
 
-    $Domain = $Credscredential.Username.Split('@')[1]
     node localhost
     {
         EXOHostedOutboundSpamFilterRule 'ConfigureHostedOutboundSpamFilterRule'