+++ Modules/Microsoft365DSC/Examples/Resources/EXOQuarantinePolicy/1-Create.ps1
@@ -18,7 +18,6 @@ Configuration Example
         [System.String]
         $CertificateThumbprint
     )
-    $OrganizationName = $Credscredential.UserName.Split('@')[1]
     Import-DscResource -ModuleName Microsoft365DSC
 
     node localhost